+++ cluster/lib/src/config.dart
@@ -13,35 +13,65 @@ class ConfigNode {
   String toString() => 'adress=$host:$port, id=$id';
 }
 
-class Config {
-  final List<ConfigNode> seedNodes;
-  final ConfigNode localNode;
-  final int workers;
-  final String secret;
+class ClusterConfig {
+  final List<ConfigNode> nodes;
+  final Map<String, int> tags;
   final int timeout;
+  final String secret;
 
-  Config(
-    this.seedNodes,
-    this.localNode,
-    this.workers,
-    this.secret,
+  ClusterConfig(
+    this.nodes,
+    this.tags,
     this.timeout,
+    this.secret,
   );
 }
 
-Future<Config> readConfigFromYaml(String filename) async {
+class NodeConfig {
+  final ConfigNode node;
+  final int workers;
+  final List<String> tags;
+
+  NodeConfig(this.node, this.workers, this.tags);
+}
+
+Future<ClusterConfig> readClusterConfigFromYaml(String filename) async {
   final yaml = await File(filename).readAsString();
   final YamlMap config = loadYaml(yaml);
 
-  return Config(
-    (config['seedNodes'] as YamlList).map((e) => ConfigNode(e['host'], e['port'], e['id'])).toList(),
+  final tags = <String, int>{};
+  if (config['tags'] != null) {
+    for (final tag in config['tags'].entries) {
+      tags[tag.key] = tag.value;
+    }
+  }
+
+  return ClusterConfig(
+    (config['nodes'] as YamlList).map((e) => ConfigNode(e['host'], e['port'], e['id'])).toList(),
+    tags,
+    config['timeout'],
+    config['secret'],
+  );
+}
+
+Future<NodeConfig> readNodeConfigFromYaml(String filename) async {
+  final yaml = await File(filename).readAsString();
+  final YamlMap config = loadYaml(yaml);
+
+  final tags = <String>[];
+  if (config['tags'] != null) {
+    for (final tag in config['tags']) {
+      tags.add(tag);
+    }
+  }
+
+  return NodeConfig(
     ConfigNode(
-      config['localNode']['host'],
-      config['localNode']['port'],
-      config['localNode']['id'],
+      config['node']['host'],
+      config['node']['port'],
+      config['node']['id'],
     ),
     config['workers'],
-    config['secret'],
-    config['timeout'],
+    tags,
   );
 }