+++ cluster/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 0.2.0
 
 - BREAKING CHANGE: based on version 0.5.0 of package `actor_system`.
+- BREAKING CHANGE: Splitted configuration into separate parts for cluster and node.
+- Cluster nodes can be configured with tags and for each tag the number of required nodes can be configured. If configured, the cluster will only start, if all required tags are available.
 
 ## 0.1.0